requirements.txt
streamlit>=1.28.0
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
numpy>=1.23.0
pandas>=1.5.0
matplotlib>=3.6.0
seaborn>=0.12.0
pillow>=9.0.0
opencv-python>=4.7.0
scikit-learn>=1.2.0
kagglehub>=0.2.0
